fully removed TabletOverlay and integrated functionality into InterfaceController
