+++ Assets/Resources/Prefabs/SnapshotManager.prefab
@@ -10,6 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 7125839906098874224}
   - component: {fileID: 7125839906098874227}
+  - component: {fileID: 2292133035155131279}
   m_Layer: 0
   m_Name: SnapshotManager
   m_TagString: Untagged
@@ -44,13 +45,23 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b42ae5e7ad3aeb245b9a0974baf59071, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  interfaceController: {fileID: 0}
   tracker: {fileID: 0}
-  tabletOverlay: {fileID: 0}
-  main: {fileID: 0}
-  host: {fileID: 0}
   trackedCamera: {fileID: 0}
   snapshotPrefab: {fileID: 5156912489491877825, guid: 459146a4ea9ca8943a5ace18241f04e7,
     type: 3}
   originPlanePrefab: {fileID: 3348506593271230907, guid: 64e19bd99f1637d4b8c9751fd20c172b,
     type: 3}
   invalidTexture: {fileID: 2800000, guid: 3423b0cb384f6d04cb6c8815c52b2ba7, type: 3}
+--- !u!114 &2292133035155131279
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7125839906098874225}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e093bb2fefac40cea8612dbb2e1a30b0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
