+++ Assets/Resources/Prefabs/TabletOverlay.prefab
@@ -699,7 +699,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1094577159368245007}
-  - component: {fileID: 1633451836285392722}
   m_Layer: 0
   m_Name: TabletOverlay
   m_TagString: Untagged
@@ -728,19 +727,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1633451836285392722
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8586341857598450562}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 306ab6531b4447ff89e052ef4767b24f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  main: {fileID: 4105797930155061690}
 --- !u!1 &8902031708148494158
 GameObject:
   m_ObjectHideFlags: 0
