+++ Assets/Scenes/Host.unity
@@ -465,18 +465,6 @@ GameObject:
     type: 3}
   m_PrefabInstance: {fileID: 2899917703023114335}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &550957534 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 4903954811414918948, guid: a4c392a7b279f8044a295d9c90e86a98,
-    type: 3}
-  m_PrefabInstance: {fileID: 2899917703023114335}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 306ab6531b4447ff89e052ef4767b24f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &553075908
 GameObject:
   m_ObjectHideFlags: 0
@@ -2882,6 +2870,11 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 2618122019900176585, guid: a4c392a7b279f8044a295d9c90e86a98,
+        type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: 81e6d19cc3d17504eaab5d8436938132, type: 2}
     - target: {fileID: 2768085287744816464, guid: a4c392a7b279f8044a295d9c90e86a98,
         type: 3}
       propertyPath: m_Name
@@ -3169,12 +3162,17 @@ PrefabInstance:
         type: 3}
       propertyPath: tabletOverlay
       value: 
-      objectReference: {fileID: 550957534}
+      objectReference: {fileID: 0}
     - target: {fileID: 7125839906098874227, guid: 481fa3b9691bdfa4cb115414591b3281,
         type: 3}
       propertyPath: trackedCamera
       value: 
       objectReference: {fileID: 406083497}
+    - target: {fileID: 7125839906098874227, guid: 481fa3b9691bdfa4cb115414591b3281,
+        type: 3}
+      propertyPath: interfaceController
+      value: 
+      objectReference: {fileID: 8142987761058685587}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 481fa3b9691bdfa4cb115414591b3281, type: 3}
 --- !u!1001 &8142987761058685586
@@ -3189,6 +3187,11 @@ PrefabInstance:
       propertyPath: hud
       value: 
       objectReference: {fileID: 998620415}
+    - target: {fileID: 1817938359801969131, guid: 75ed3f0ab7eb6df41ab594118d5e5b9d,
+        type: 3}
+      propertyPath: main
+      value: 
+      objectReference: {fileID: 571167376}
     - target: {fileID: 1817938359801969131, guid: 75ed3f0ab7eb6df41ab594118d5e5b9d,
         type: 3}
       propertyPath: centerText
