--- Assets/Scripts/Snapshots/TabletOverlay.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 306ab6531b4447ff89e052ef4767b24f
-timeCreated: 1695105038
